v72.5: Code quality fixes from debug reports
Fixed:
- Removed 150+ lines of dead code in admin-documents.php (unreachable code after return)
- Fixed float to int type casting for total_pages in admin-bookings.php
- Fixed int to string type casting for esc_attr() in admin.php

Code Quality Analysis:
- Reviewed 3,355+ issues from PHPStan, Psalm, PHPMD, and PHPCS
- Most issues were false positives due to WordPress architecture
- Defensive return statements after wp_send_json_*() are intentional

Files modified:
- yolo-yacht-search.php (version bump)
- admin/class-yolo-ys-admin-documents.php
- admin/class-yolo-ys-admin-bookings.php
- admin/class-yolo-ys-admin.php
- CHANGELOG.md

+++ yolo-yacht-search/admin/class-yolo-ys-admin-bookings.php
@@ -237,7 +237,7 @@ class YOLO_YS_Admin_Bookings extends WP_List_Table {
         $this->set_pagination_args(array(
             'total_items' => $total_items,
             'per_page'    => $per_page,
-            'total_pages' => ceil($total_items / $per_page)
+            'total_pages' => (int) ceil($total_items / $per_page)
         ));
 
         // Set columns
